Got frustated. Might delete later. IDK. LOL
WIP with elasticsearch ingest type node.
For that trying to set up elasticsearch in more of a production mode,
trying to set up a 'cluster'. Didn't get how to do that,
because changing discovery.type to 'multi-node' or 'multiple-node'
didn't yield any results.
I couldn't find what are valid values for discovery.type.
Also pulled out ingress'es and most services to separate file,
to be easier to manage, hopefully.