Updates for Gitlab config
Take Gitlab.rb from configmap
Allow Traefik to pass through acme-challenge.
https://gist.github.com/micw/67faf5cd3d4a6f64568ca2bb9a051230 helped a
lot to get it working.
That was required, because Traefik intercepts the
/.well-known/acme-challenge url by default.
Thus not allowing Gitlab's nginx to complete the acme challenge, as
Traefik tried to do it instead.
All's good that's fixed.
Removed Traefik global http to https redirect, to allow it to I guess.