Elastic APM RUM, and trying PersistentVolume
Persistent volumes didn't give the benefit I wanted.
Might be the local type is just not what I need.
Still have to make sure that permissions are set for user id 1000,
to have elasticsearch working properly, even with PV.
But meh, that's not something that I need urgently for now.