Playing around with ECK
So far I seem to prefer the previous yml templates better.
Feels like faster that way for small tests,
but these CRD's with the operator make sure that
another instance of Kibana is up, before terminating old one,
so that's nice.
Basically just an exercise in setting up Elastic this way.