Testing securing Elastic http comms
Even though the documentation says to set up certs on each node
It seems to work also just setting it on node, that Kibana connects to.
So to get alarms on Kibana, one can set up separate Elasticsearch node
with cert, and have another Elasticsearch node for ingestion without certs.
Not good approach, but might allow for easier migration to secure setup.
Instead of needing to update all beats, to use the cert immediately,
separate ingest node might be created with cert and beats updated
with possibly less stress, to point to secured Elasticsearch node.

Best approach would be to generate CSR and get proper cert for this.
Will see if I can do that with LetsEncrypt eventually.